requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
Django==5.0.4
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21
drf-spectacular==0.27.2
flake8==7.0.0
flake8-docstrings==1.7.0
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sentry-sdk==1.45.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
